added new component.  renamed some stuff.
