+++ src/App.jsx
@@ -4,6 +4,7 @@ import styled from "styled-components";
 import { HashRouter, Route } from "react-router-dom";
 import Home from "./Home.jsx";
 import About from "./About.jsx";
+import Explore from "./Explore.jsx";
 import Nav from "./Nav.jsx";
 
 class App extends Component {
@@ -16,6 +17,7 @@ class App extends Component {
             <div>
               <Route exact path="/" component={Home} />
               <Route exact path="/about" component={About} />
+              <Route exact path="/explore" component={Explore} />
             </div>
           </div>
         </HashRouter>
